requirements.txt
alembic==1.7.3
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
Flask==2.0.1
Flask-JWT-Extended==4.3.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
google-auth==2.3.0
greenlet==1.1.2
httplib2==0.20.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jwt==1.3.0
Mako==1.1.5
MarkupSafe==2.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==2.0.1
